requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-auth==2.36.0
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
